restructuring the prepare function
