+++ .eslintrc
@@ -48,7 +48,8 @@
 		"@typescript-eslint/restrict-plus-operands": "error",
 		"@typescript-eslint/type-annotation-spacing": "error",
 		"@typescript-eslint/unbound-method": "error",
-		"i18n-text/no-en": "off"
+		"i18n-text/no-en": "off",
+		"prefer-const": [0, {"destructuring": "all"}]
 	},
 	"env": {
 		"node": true,
