+++ src/main.ts
@@ -1,8 +1,8 @@
 import * as core from '@actions/core'
-import * as github from '@actions/github'
 import { exec } from '@actions/exec'
-import { comment as githubComment } from './commentToPullRequest'
-import { execSurgeCommand, formatImage } from './helpers'
+import { comment, fail } from './actions'
+import { execCommand } from './helpers/execCommand'
+import { formatImage } from './helpers/formatImage'
 import {
 	vercelDeploy,
 	vercelRemoveProjectDeploy,
@@ -10,178 +10,27 @@ import {
 } from './tenants/vercel'
 import prepare from './previewPipeline/prepare'
 
-function getGitCommitSha(): string {
-	const { payload } = github.context
-	const gitCommitSha =
-		payload.after ||
-		payload?.pull_request?.head?.sha ||
-		payload?.workflow_run?.head_sha
-
-	return gitCommitSha
-}
-
-async function getPullRequestNumber(): Promise<number | undefined> {
-	const token = core.getInput('github_token', { required: true })
-	const octokit = github.getOctokit(token)
-	const { payload } = github.context
-	const gitCommitSha = getGitCommitSha()
-	const prNumberExists = payload.number && payload.pull_request
-
-	if (prNumberExists) {
-		return Number(payload.number)
-	}
-
-	if (!prNumberExists) {
-		const result =
-			await octokit.rest.repos.listPullRequestsAssociatedWithCommit({
-				owner: github.context.repo.owner,
-				repo: github.context.repo.repo,
-				commit_sha: gitCommitSha,
-			})
-		const pr = result.data.length > 0 && result.data[0]
-		core.debug('listPullRequestsAssociatedWithCommit')
-		core.debug(JSON.stringify(pr, null, 2))
-		const prNumber = pr ? Number(pr.number) : undefined
-		return prNumber
-	}
-}
-
-async function comment(message: string): Promise<void> {
-	const { job, payload } = github.context
-	const prOwner = payload.pull_request?.owner
-	const fromForkedRepo = prOwner === github.context.repo.owner
-	const token = core.getInput('github_token', { required: true })
-	const octokit = github.getOctokit(token)
-	const prNumber = await getPullRequestNumber()
-
-	if (fromForkedRepo) {
-		return
-	}
-
-	githubComment({
-		repo: github.context.repo,
-		number: Number(prNumber),
-		message,
-		octokit,
-		header: job,
-	})
-}
-
-async function generateLogUrl(): Promise<string> {
-	const token = core.getInput('github_token', { required: true })
-	const octokit = github.getOctokit(token)
-	const { job } = github.context
-
-	const gitCommitSha = getGitCommitSha()
-
-	let data
-	try {
-		const result = await octokit.rest.checks.listForRef({
-			owner: github.context.repo.owner,
-			repo: github.context.repo.repo,
-			ref: gitCommitSha,
-		})
-		data = result.data
-	} catch (err) {
-		core.info('generateLogUrl error')
-		await fail(err)
-		return ''
-	}
-
-	core.debug(JSON.stringify(data?.check_runs, null, 2))
-
-	let checkRunId
-	if (data?.check_runs?.length >= 0) {
-		const checkRun = data?.check_runs?.find((item) => item.name === job)
-		checkRunId = checkRun?.id
-	}
-
-	const buildingLogUrl = checkRunId
-		? `https://github.com/${github.context.repo.owner}/${github.context.repo.repo}/runs/${checkRunId}`
-		: `https://github.com/${github.context.repo.owner}/${github.context.repo.repo}/actions/runs/${github.context.runId}`
-
-	return buildingLogUrl
-}
-
-async function fail(err: Error) {
-	core.info('error message:')
-	core.info(JSON.stringify(err, null, 2))
-	const repoOwner = github.context.repo.owner
-	const repoName = github.context.repo.repo
-	const repoId = github.context.runId
-	const buildLogsUrl = `https://github.com/${repoOwner}/${repoName}/actions/runs/${repoId}`
-	const buildingLogUrl = await generateLogUrl()
-
-	const gitCommitSha = getGitCommitSha()
-
-	const image = formatImage({
-		buildingLogUrl,
-		imageUrl:
-			'https://user-images.githubusercontent.com/507615/90250824-4e066700-de6f-11ea-8230-600ecc3d6a6b.png',
-	})
-
-	await comment(
-		`😭 Deploy PR Preview ${gitCommitSha} failed. [Build logs](${buildLogsUrl}) \n ${image}`
-	)
-
-	const failOnError = !!(
-		core.getInput('failOnError') || process.env.FAIL_ON__ERROR
-	)
-
-	if (failOnError) {
-		core.setFailed(err.message)
-	}
-}
-
 async function main() {
 	const {
 		surgeToken,
-		previewUrl,
 		previewPath,
 		distFolder,
 		teardown,
-		prNumber,
-		jobContext,
 		payloadContext,
 		gitCommitSha,
-	} = await prepare({ getPullRequestNumber, getGitCommitSha })
-
-	if (!prNumber) {
-		core.info(`😢 No related PR found, skip it.`)
-		return
-	}
-	core.info(`Find PR number: ${prNumber}`)
-
-	const repoOwner = github.context.repo.owner.replace(/\./g, '-')
-	const repoName = github.context.repo.repo.replace(/\./g, '-')
-	const url = previewUrl
-		.replace('{{repoOwner}}', repoOwner)
-		.replace('{{repoName}}', repoName)
-		.replace('{{job}}', jobContext)
-		.replace('{{prNumber}}', `${prNumber}`)
-		.concat('.surge.sh')
-
-	const outputUrl = url.concat(previewPath)
-
-	core.setOutput('preview_url', outputUrl)
-
-	const buildingLogUrl = await generateLogUrl()
-
-	core.debug(`teardown enabled?: ${teardown}`)
-	core.debug(`event action?: ${payloadContext.action}`)
-
-	// Vercel
-	core.info('Init config vercel')
-	const vercelToken = core.getInput('vercel_token')
-	let deploymentUrlVercel = ''
-	// Vercel
+		mountedUrl,
+		outputUrl,
+		buildingLogUrl,
+		configVercel,
+	} = await prepare()
+	let { vercelToken, deploymentUrlVercel } = configVercel
 
 	if (teardown && payloadContext.action === 'closed') {
 		try {
-			core.info(`Teardown: ${url}`)
+			core.info(`Teardown: ${mountedUrl}`)
 			core.setSecret(surgeToken)
-			await execSurgeCommand({
-				command: ['surge', 'teardown', url, `--token`, surgeToken],
+			await execCommand({
+				command: ['surge', 'teardown', mountedUrl, `--token`, surgeToken],
 			})
 
 			const image = formatImage({
@@ -225,7 +74,7 @@ async function main() {
 		}
 		const duration = (Date.now() - startTime) / 1000
 		core.info(`Build time: ${duration} seconds`)
-		core.info(`Deploy to ${url}`)
+		core.info(`Deploy to ${mountedUrl}`)
 		core.setSecret(surgeToken)
 		const image = formatImage({
 			buildingLogUrl,
@@ -239,8 +88,8 @@ async function main() {
 		}
 		// Vercel
 
-		await execSurgeCommand({
-			command: ['surge', `./${distFolder}`, url, `--token`, surgeToken],
+		await execCommand({
+			command: ['surge', `./${distFolder}`, mountedUrl, `--token`, surgeToken],
 		})
 
 		await comment(`
